+++ app.js
@@ -19,7 +19,8 @@ app.use((req, res) => {
 })
 
 app.use((err, req, res, next) => {
-  res.status(err.status).json({ message: err.message })
+  const { status = 500, message = "Server error" } = err;
+  res.status(status).json({ message: message })
 })
 
 module.exports = app
