+++ models/contacts/filePath.js
@@ -1,0 +1,4 @@
+const path = require("path");
+const filePath = path.join(__dirname, "contacts.json");
+
+module.exports = filePath;
